Refactor HboGo and Youtube to use Channel interface
Not quite complete yet; the Youtube tests fail...
